cloud_functions/requirements.txt
functions-framework==3.3.0
pandas
google-api-python-client
google-auth
google-auth-oauthlib 
google-cloud-storage
google-cloud-pubsub
requests
google-cloud-secret-manager
flask
fsspec
gcsfs
pyyaml
